nlp-service/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic-settings>=2.2.0
openai-whisper>=20231117
transformers>=4.40.0
torch>=2.2.0
python-multipart>=0.0.9
ollama>=0.2.0
pydantic>=2.0.0
